+++ testUtils/createEastrc.js
@@ -2,7 +2,7 @@
 
 const pathUtils = require('path');
 const _ = require('underscore');
-const fse = require('fs-extra');
+const fs = require('fs');
 
 // allow to enable config loading by env var, useful for integration
 // testing with different adapters
@@ -30,13 +30,20 @@ module.exports = (params) => {
 			if (loadConfig) {
 				const cwdConfigPath = pathUtils.resolve('.eastrc');
 
-				return fse.readJSON(cwdConfigPath);
+				return fs.promises.readFile(cwdConfigPath, 'utf-8');
 			}
 		})
-		.then((cwdConfigParams) => {
-			return fse.writeJSON(
+		.then((cwdConfigParamsText) => {
+			let cwdConfigParams;
+			if (cwdConfigParamsText) {
+				cwdConfigParams = JSON.parse(cwdConfigParamsText);
+			} else {
+				cwdConfigParams = {};
+			}
+			return fs.promises.writeFile(
 				configPath,
-				_(configParams).defaults(cwdConfigParams)
+				JSON.stringify(_(configParams).defaults(cwdConfigParams)),
+				'utf-8'
 			);
 		})
 		.then(() => configPath);